+++ lib/ensql.rb
@@ -25,12 +25,10 @@ require_relative "ensql/sql"
 #     Ensql.sql('select * from users where id = %{id}', id: 1).first_row # => { "id" => 1, "email" => "test@example.com" }
 #
 module Ensql
-
   class << self
-
     # (see SQL)
     # @return [Ensql::SQL] SQL statement
-    def sql(sql, params={})
+    def sql(sql, params = {})
       SQL.new(sql, params)
     end
 
@@ -43,7 +41,7 @@ module Ensql
     #   Ensql.sql_path = Rails.root.join('app/queries')
     #
     def sql_path
-      @sql_path ||= 'sql'
+      @sql_path ||= "sql"
     end
     attr_writer :sql_path
 
@@ -58,7 +56,7 @@ module Ensql
     #   Ensql.load_sql('users/activity', report_params)
     #   Ensql.load_sql(:upsert_users, imported_users_attrs)
     #
-    def load_sql(name, params={})
+    def load_sql(name, params = {})
       path = File.join(sql_path, "#{name}.sql")
       SQL.new(File.read(path), params, name)
     end
@@ -71,9 +69,8 @@ module Ensql
     #   Ensql.run("DELETE FROM users WHERE id = %{id}", id: user.id)
     #   Ensql.run("ALTER TABLE test RENAME TO old_test")
     #
-    def run(sql, params={})
+    def run(sql, params = {})
       SQL.new(sql, params).run
     end
-
   end
 end